fix(): refactoring menu as component
